pyproject.toml
[project]
name = "congressional-hackathon-2025"
version = "0.1.0"
description = "App to compare passed laws against sponsor campaign objectives using LLM analysis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "streamlit>=1.39.0",
    "openai>=1.58.0",
    "anthropic>=0.40.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "urllib3>=2.0.0",
    "lxml>=5.0.0",
    "duckduckgo-search>=6.0.0",
    "ddgs>=9.6.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.7",
]
